Added support for obtaining Ec token, reorganzed source hierarchy and seperated services, minor misc tweaks.

+++ include/util.hpp
@@ -7,11 +7,12 @@
 #pragma once
 
 #include <string>
+#include <cstdint>
 #include <citrus/core.hpp>
 
 using namespace std;
 
 class Util {
 public:
-    static string base64Encode(u8 const* bytesToEnc, u32 bufLen);
+    static string base64Encode(u8 const* bytesToEnc, size_t bufLen);
 };
